apps/scraper/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
aiosqlite>=0.20.0
pyyaml>=6.0
